+++ CHANGELOG.md
@@ -5,3 +5,4 @@ This file is used to list changes made in each version of the GitHub cookbook.
 ## Unreleased
 
 - Sous-Chefs adoption
+- Migrate to modern custom resources